fix unit test.
