+++ cmd/kafka-consumer/main.go
@@ -34,6 +34,7 @@ import (
 	"time"
 
 	"github.com/IBM/sarama"
+	"github.com/edwingeng/deque"
 	"github.com/google/uuid"
 	cerror "github.com/pingcap/errors"
 	"github.com/pingcap/log"
@@ -43,7 +44,6 @@ import (
 	eventsinkfactory "github.com/pingcap/tiflow/cdc/sink/dmlsink/factory"
 	"github.com/pingcap/tiflow/cdc/sink/dmlsink/mq/dispatcher"
 	"github.com/pingcap/tiflow/cdc/sink/tablesink"
-	"github.com/pingcap/tiflow/cmd/kafka-consumer/flowcontrol"
 	cmdUtil "github.com/pingcap/tiflow/pkg/cmd/util"
 	"github.com/pingcap/tiflow/pkg/config"
 	"github.com/pingcap/tiflow/pkg/filter"
@@ -310,7 +310,7 @@ func main() {
 	go func() {
 		defer wg.Done()
 		for {
-			// `Consume` should be called inside an infinite loop, when a
+			// `consume` should be called inside an infinite loop, when a
 			// server-side rebalance happens, the consumer session will need to be
 			// recreated to get the new claims
 			if err := client.Consume(ctx, strings.Split(consumerOption.topic, ","), consumer); err != nil {
@@ -432,7 +432,7 @@ type partitionSinks struct {
 	// resolvedTs record the maximum timestamp of the received event
 	resolvedTs uint64
 
-	flowController *flowcontrol.FlowController
+	flowController *flowController
 }
 
 // Consumer represents a Sarama consumer group consumer
@@ -518,7 +518,7 @@ func NewConsumer(ctx context.Context, o *consumerOption) (*Consumer, error) {
 	memoryQuotaPerPartition := defaultMemoryQuotaInBytes / int(o.partitionNum)
 	for i := 0; i < int(o.partitionNum); i++ {
 		c.sinks[i] = &partitionSinks{
-			flowController: flowcontrol.NewFlowController(uint64(memoryQuotaPerPartition)),
+			flowController: newFlowController(uint64(memoryQuotaPerPartition)),
 		}
 	}
 
@@ -729,9 +729,9 @@ func (c *Consumer) ConsumeClaim(session sarama.ConsumerGroupSession, claim saram
 				// todo: mark the offset after the DDL is fully synced to the downstream mysql.
 				session.MarkMessage(message, "")
 
-				err = sink.flowController.Consume(row.CommitTs, uint64(row.ApproximateBytes()))
+				err = sink.flowController.consume(row.CommitTs, uint64(row.ApproximateBytes()))
 				if err != nil {
-					if errors.Is(err, flowcontrol.ErrorFlowControllerAborted) {
+					if errors.Is(err, errFlowControllerAborted) {
 						log.Info("flow control aborted")
 						return nil
 					}
@@ -951,7 +951,7 @@ func syncFlushRowChangedEvents(ctx context.Context, sink *partitionSinks, resolv
 			return true
 		})
 		if flushedResolvedTs {
-			sink.flowController.Release(resolvedTs)
+			sink.flowController.release(resolvedTs)
 			return nil
 		}
 	}
@@ -997,4 +997,141 @@ func openDB(ctx context.Context, dsn string) (*sql.DB, error) {
 	}
 	log.Info("open db success", zap.String("dsn", dsn))
 	return db, nil
+}
+
+var (
+	errFlowControllerLargerThanQuota = errors.New("flow controller request memory larger than quota")
+	errFlowControllerAborted         = errors.New("flow controller aborted")
+)
+
+type memoryQuota struct {
+	quota uint64 // should not be changed once initialized
+
+	isAborted atomic.Bool
+
+	consumed struct {
+		sync.Mutex
+		bytes uint64
+	}
+
+	consumedCond *sync.Cond
+}
+
+// newMemoryQuota creates a new memoryQuota
+// quota: max advised memory consumption in bytes.
+func newMemoryQuota(quota uint64) *memoryQuota {
+	ret := &memoryQuota{
+		quota: quota,
+	}
+
+	ret.consumedCond = sync.NewCond(&ret.consumed)
+	return ret
+}
+
+// consumeWithBlocking is called when a hard-limit is needed. The method will
+// block until enough memory has been freed up by release.
+// blockCallBack will be called if the function will block.
+// Should be used with care to prevent deadlock.
+func (c *memoryQuota) consumeWithBlocking(nBytes uint64) error {
+	if nBytes >= c.quota {
+		return errFlowControllerLargerThanQuota
+	}
+
+	c.consumed.Lock()
+	defer c.consumed.Unlock()
+
+	for {
+		if c.isAborted.Load() {
+			return errFlowControllerAborted
+		}
+
+		if c.consumed.bytes+nBytes < c.quota {
+			break
+		}
+		c.consumedCond.Wait()
+	}
+
+	c.consumed.bytes += nBytes
+	return nil
+}
+
+// release is called when a chuck of memory is done being used.
+func (c *memoryQuota) release(nBytes uint64) {
+	c.consumed.Lock()
+
+	if c.consumed.bytes < nBytes {
+		c.consumed.Unlock()
+		log.Panic("memoryQuota: releasing more than consumed, report a bug",
+			zap.Uint64("consumed", c.consumed.bytes),
+			zap.Uint64("released", nBytes))
+	}
+
+	c.consumed.bytes -= nBytes
+	if c.consumed.bytes < c.quota {
+		c.consumed.Unlock()
+		c.consumedCond.Signal()
+		return
+	}
+
+	c.consumed.Unlock()
+}
+
+type flowController struct {
+	memoryQuota *memoryQuota
+
+	queueMu struct {
+		sync.Mutex
+		queue deque.Deque
+	}
+}
+
+type entry struct {
+	commitTs uint64
+	size     uint64
+}
+
+func newFlowController(quota uint64) *flowController {
+	return &flowController{
+		memoryQuota: newMemoryQuota(quota),
+		queueMu: struct {
+			sync.Mutex
+			queue deque.Deque
+		}{
+			queue: deque.NewDeque(),
+		},
+	}
+}
+
+func (c *flowController) consume(commitTs uint64, size uint64) error {
+	err := c.memoryQuota.consumeWithBlocking(size)
+	if err != nil {
+		return cerror.Trace(err)
+	}
+
+	c.queueMu.Lock()
+	defer c.queueMu.Unlock()
+
+	c.queueMu.queue.PushBack(&entry{
+		commitTs: commitTs,
+		size:     size,
+	})
+
+	return nil
+}
+
+func (c *flowController) release(resolvedTs uint64) {
+	var nBytesToRelease uint64
+
+	c.queueMu.Lock()
+	for c.queueMu.queue.Len() > 0 {
+		if peeked := c.queueMu.queue.Front().(*entry); peeked.commitTs <= resolvedTs {
+			nBytesToRelease += peeked.size
+			c.queueMu.queue.PopFront()
+		} else {
+			break
+		}
+	}
+	c.queueMu.Unlock()
+
+	c.memoryQuota.release(nBytesToRelease)
 }
